Improve configuration cache compatibility
